server/requirements.txt
python-dotenv
supabase
langchain
langchain-community
langchain-groq
flask
flask_cors
pymupdf
langchain-cohere
pymupdf
langchain-neo4j
rapidfuzz
pandas
ragas
datasets
openpyxl
langchain>=0.1.0
langchain-community>=0.0.1
langchain-core>=0.0.1
langchain-groq>=0.0.1
supabase>=1.0.0
python-dotenv>=1.0.0
cohere>=4.0.0
groq>=0.1.0
neo4j>=5.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
sentence-transformers>=2.2.2
livekit-agents[groq]~=1.0
livekit-agents[silero]~=1.0
fastapi
uvicorn
python-multipart 
pydub 
